Remove shelve endpoint.
